kernel: Fix scheduler tracing formats and priority field type.
* Use format macros for printing to fix priting on 64 bit.
* Use int32 for all priority fields, some of them were truncated.

+++ src/system/kernel/scheduler/scheduler_tracing.h
@@ -73,7 +73,7 @@ public:
 	virtual const char* Name() const;
 
 private:
-	uint8				fPriority;
+	int32				fPriority;
 };
 
 
@@ -114,7 +114,7 @@ private:
 	thread_id			fPreviousID;
 	int32				fCPU;
 	char*				fName;
-	uint8				fPriority;
+	int32				fPriority;
 	uint8				fPreviousState;
 	uint16				fPreviousWaitObjectType;
 	union {
